# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Job {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  capacity: number;
  activeCount: number;
  waitlistCount: number;
  createdAt: string;
}

export interface CreateJobBody {
  title: string;
  description?: string;
  /** @minimum 1 */
  capacity: number;
}

export type ApplicationEntryStatus =
  (typeof ApplicationEntryStatus)[keyof typeof ApplicationEntryStatus];

export const ApplicationEntryStatus = {
  ACTIVE: "ACTIVE",
  WAITLIST: "WAITLIST",
  PENDING_ACKNOWLEDGMENT: "PENDING_ACKNOWLEDGMENT",
  INACTIVE: "INACTIVE",
} as const;

export interface ApplicationEntry {
  applicationId: number;
  applicantId: number;
  applicantName: string;
  applicantEmail: string;
  status: ApplicationEntryStatus;
  /** @nullable */
  queuePosition?: number | null;
  appliedAt: string;
  /** @nullable */
  promotedAt?: string | null;
  /** @nullable */
  acknowledgeDeadline?: string | null;
}

export interface JobWithPipeline {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  capacity: number;
  activeApplicants: ApplicationEntry[];
  waitlistApplicants: ApplicationEntry[];
  createdAt: string;
}

export interface Applicant {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface CreateApplicantBody {
  name: string;
  email: string;
}

export interface ApplyBody {
  applicantId: number;
  jobId: number;
}

export interface WithdrawBody {
  applicationId: number;
}

export interface AcknowledgeBody {
  applicationId: number;
}

export type ApplicationResultStatus =
  (typeof ApplicationResultStatus)[keyof typeof ApplicationResultStatus];

export const ApplicationResultStatus = {
  ACTIVE: "ACTIVE",
  WAITLIST: "WAITLIST",
  PENDING_ACKNOWLEDGMENT: "PENDING_ACKNOWLEDGMENT",
  INACTIVE: "INACTIVE",
} as const;

export interface ApplicationResult {
  applicationId: number;
  applicantId: number;
  jobId: number;
  status: ApplicationResultStatus;
  /** @nullable */
  queuePosition?: number | null;
  message?: string;
}

export type ApplicationStatusEntryStatus =
  (typeof ApplicationStatusEntryStatus)[keyof typeof ApplicationStatusEntryStatus];

export const ApplicationStatusEntryStatus = {
  ACTIVE: "ACTIVE",
  WAITLIST: "WAITLIST",
  PENDING_ACKNOWLEDGMENT: "PENDING_ACKNOWLEDGMENT",
  INACTIVE: "INACTIVE",
} as const;

export interface ApplicationStatusEntry {
  applicationId: number;
  jobId: number;
  jobTitle: string;
  status: ApplicationStatusEntryStatus;
  /** @nullable */
  queuePosition?: number | null;
  appliedAt: string;
  /** @nullable */
  promotedAt?: string | null;
  /** @nullable */
  acknowledgeDeadline?: string | null;
  timeInCurrentStateSeconds?: number;
}

export interface ApplicantStatus {
  applicantId: number;
  applicantName: string;
  applicantEmail: string;
  applications: ApplicationStatusEntry[];
}

export interface QueueEntry {
  applicationId: number;
  applicantId: number;
  applicantName: string;
  position: number;
  appliedAt: string;
  penaltyCount: number;
}

export interface JobQueue {
  jobId: number;
  jobTitle: string;
  capacity: number;
  activeCount: number;
  waitlistEntries: QueueEntry[];
}

export type AuditLogMetadata = { [key: string]: unknown };

export interface AuditLog {
  id: number;
  applicationId: number;
  eventType: string;
  /** @nullable */
  fromStatus?: string | null;
  toStatus: string;
  metadata?: AuditLogMetadata;
  createdAt: string;
}

export interface PipelineSummary {
  jobId: number;
  jobTitle: string;
  capacity: number;
  activeCount: number;
  waitlistCount: number;
  inactiveCount: number;
  totalApplications: number;
  /** @nullable */
  avgTimeToPromotionSeconds?: number | null;
  /** @nullable */
  avgAcknowledgmentTimeSeconds?: number | null;
  decayEvents: number;
}

export interface ReplayEntry {
  applicationId: number;
  applicantId: number;
  applicantName: string;
  status: string;
}

export interface PipelineReplay {
  jobId: number;
  asOf: string;
  activeApplicants: ReplayEntry[];
  waitlistApplicants: ReplayEntry[];
  events: AuditLog[];
}

export type ReplayPipelineParams = {
  asOf?: string;
};