# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Self-Moving Hiring Pipeline API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applicants:
    get:
      operationId: listApplicants
      tags: [applicants]
      summary: List all applicants
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Applicant"
    post:
      operationId: createApplicant
      tags: [applicants]
      summary: Register a new applicant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicantBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applicants/{applicantId}:
    get:
      operationId: getApplicant
      tags: [applicants]
      summary: Get applicant details
      parameters:
        - name: applicantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /apply:
    post:
      operationId: applyToJob
      tags: [applications]
      summary: Apply to a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyBody"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResult"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /withdraw:
    post:
      operationId: withdrawApplication
      tags: [applications]
      summary: Withdraw an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "200":
          description: Withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResult"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /acknowledge:
    post:
      operationId: acknowledgePromotion
      tags: [applications]
      summary: Acknowledge a promotion within the time window
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcknowledgeBody"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResult"
        "400":
          description: Bad Request or window expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /status/{applicantId}:
    get:
      operationId: getApplicantStatus
      tags: [applicants]
      summary: Get an applicant's current status across all jobs
      parameters:
        - name: applicantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantStatus"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /timeline/{applicantId}:
    get:
      operationId: getApplicantTimeline
      tags: [applicants]
      summary: Get the full event timeline for an applicant
      parameters:
        - name: applicantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    Applicant:
      type: object
      required:
        - id
        - name
        - email
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateApplicantBody:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
    ApplyBody:
      type: object
      required:
        - applicantId
        - jobId
      properties:
        applicantId:
          type: integer
        jobId:
          type: integer
    WithdrawBody:
      type: object
      required:
        - applicationId
      properties:
        applicationId:
          type: integer
    AcknowledgeBody:
      type: object
      required:
        - applicationId
      properties:
        applicationId:
          type: integer
    ApplicationResult:
      type: object
      required:
        - applicationId
        - applicantId
        - jobId
        - status
      properties:
        applicationId:
          type: integer
        applicantId:
          type: integer
        jobId:
          type: integer
        status:
          type: string
          enum: [ACTIVE, WAITLIST, PENDING_ACKNOWLEDGMENT, INACTIVE]
        queuePosition:
          type: ["integer", "null"]
        message:
          type: string
    ApplicantStatus:
      type: object
      required:
        - applicantId
        - applicantName
        - applicantEmail
        - applications
      properties:
        applicantId:
          type: integer
        applicantName:
          type: string
        applicantEmail:
          type: string
        applications:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationStatusEntry"
    ApplicationStatusEntry:
      type: object
      required:
        - applicationId
        - jobId
        - jobTitle
        - status
        - appliedAt
      properties:
        applicationId:
          type: integer
        jobId:
          type: integer
        jobTitle:
          type: string
        status:
          type: string
          enum: [ACTIVE, WAITLIST, PENDING_ACKNOWLEDGMENT, INACTIVE]
        queuePosition:
          type: ["integer", "null"]
        appliedAt:
          type: string
          format: date-time
        promotedAt:
          type: ["string", "null"]
          format: date-time
        acknowledgeDeadline:
          type: ["string", "null"]
          format: date-time
        timeInCurrentStateSeconds:
          type: integer
...
    AuditLog:
      type: object
      required:
        - id
        - applicationId
        - eventType
        - toStatus
        - createdAt
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        eventType:
          type: string
        fromStatus:
          type: ["string", "null"]
        toStatus:
          type: string
        metadata:
          type: object
          additionalProperties: {}
        createdAt:
          type: string
          format: date-time
...