# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all jobs
 */
export const ListJobsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  capacity: zod.number(),
  activeCount: zod.number(),
  waitlistCount: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListJobsResponse = zod.array(ListJobsResponseItem);

/**
 * @summary Create a new job
 */

export const CreateJobBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  capacity: zod.number().min(1),
});

/**
 * @summary Get a job with its pipeline
 */
export const GetJobParams = zod.object({
  jobId: zod.coerce.number(),
});

export const GetJobResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  capacity: zod.number(),
  activeApplicants: zod.array(
    zod.object({
      applicationId: zod.number(),
      applicantId: zod.number(),
      applicantName: zod.string(),
      applicantEmail: zod.string(),
      status: zod.enum([
        "ACTIVE",
        "WAITLIST",
        "PENDING_ACKNOWLEDGMENT",
        "INACTIVE",
      ]),
      queuePosition: zod.number().nullish(),
      appliedAt: zod.coerce.date(),
      promotedAt: zod.coerce.date().nullish(),
      acknowledgeDeadline: zod.coerce.date().nullish(),
    }),
  ),
  waitlistApplicants: zod.array(
    zod.object({
      applicationId: zod.number(),
      applicantId: zod.number(),
      applicantName: zod.string(),
      applicantEmail: zod.string(),
      status: zod.enum([
        "ACTIVE",
        "WAITLIST",
        "PENDING_ACKNOWLEDGMENT",
        "INACTIVE",
      ]),
      queuePosition: zod.number().nullish(),
      appliedAt: zod.coerce.date(),
      promotedAt: zod.coerce.date().nullish(),
      acknowledgeDeadline: zod.coerce.date().nullish(),
    }),
  ),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all applicants
 */
export const ListApplicantsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListApplicantsResponse = zod.array(ListApplicantsResponseItem);

/**
 * @summary Register a new applicant
 */
export const CreateApplicantBody = zod.object({
  name: zod.string(),
  email: zod.string(),
});

/**
 * @summary Get applicant details
 */
export const GetApplicantParams = zod.object({
  applicantId: zod.coerce.number(),
});

export const GetApplicantResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Apply to a job
 */
export const ApplyToJobBody = zod.object({
  applicantId: zod.number(),
  jobId: zod.number(),
});

/**
 * @summary Withdraw an application
 */
export const WithdrawApplicationBody = zod.object({
  applicationId: zod.number(),
});

export const WithdrawApplicationResponse = zod.object({
  applicationId: zod.number(),
  applicantId: zod.number(),
  jobId: zod.number(),
  status: zod.enum([
    "ACTIVE",
    "WAITLIST",
    "PENDING_ACKNOWLEDGMENT",
    "INACTIVE",
  ]),
  queuePosition: zod.number().nullish(),
  message: zod.string().optional(),
});

/**
 * @summary Acknowledge a promotion within the time window
 */
export const AcknowledgePromotionBody = zod.object({
  applicationId: zod.number(),
});

export const AcknowledgePromotionResponse = zod.object({
  applicationId: zod.number(),
  applicantId: zod.number(),
  jobId: zod.number(),
  status: zod.enum([
    "ACTIVE",
    "WAITLIST",
    "PENDING_ACKNOWLEDGMENT",
    "INACTIVE",
  ]),
  queuePosition: zod.number().nullish(),
  message: zod.string().optional(),
});

/**
 * @summary Get an applicant's current status across all jobs
 */
export const GetApplicantStatusParams = zod.object({
  applicantId: zod.coerce.number(),
});

export const GetApplicantStatusResponse = zod.object({
  applicantId: zod.number(),
  applicantName: zod.string(),
  applicantEmail: zod.string(),
  applications: zod.array(
    zod.object({
      applicationId: zod.number(),
      jobId: zod.number(),
      jobTitle: zod.string(),
      status: zod.enum([
        "ACTIVE",
        "WAITLIST",
        "PENDING_ACKNOWLEDGMENT",
        "INACTIVE",
      ]),
      queuePosition: zod.number().nullish(),
      appliedAt: zod.coerce.date(),
      promotedAt: zod.coerce.date().nullish(),
      acknowledgeDeadline: zod.coerce.date().nullish(),
      timeInCurrentStateSeconds: zod.number().optional(),
    }),
  ),
});

/**
 * @summary Get the full waitlist queue for a job
 */
export const GetJobQueueParams = zod.object({
  jobId: zod.coerce.number(),
});

export const GetJobQueueResponse = zod.object({
  jobId: zod.number(),
  jobTitle: zod.string(),
  capacity: zod.number(),
  activeCount: zod.number(),
  waitlistEntries: zod.array(
    zod.object({
      applicationId: zod.number(),
      applicantId: zod.number(),
      applicantName: zod.string(),
      position: zod.number(),
      appliedAt: zod.coerce.date(),
      penaltyCount: zod.number(),
    }),
  ),
});

/**
 * @summary Get the full event timeline for an applicant
 */
export const GetApplicantTimelineParams = zod.object({
  applicantId: zod.coerce.number(),
});

export const GetApplicantTimelineResponseItem = zod.object({
  id: zod.number(),
  applicationId: zod.number(),
  eventType: zod.string(),
  fromStatus: zod.string().nullish(),
  toStatus: zod.string(),
  metadata: zod.record(zod.string(), zod.unknown()).optional(),
  createdAt: zod.coerce.date(),
});
export const GetApplicantTimelineResponse = zod.array(
  GetApplicantTimelineResponseItem,
);

/**
 * @summary Get a job pipeline summary with counts and metrics
 */
export const GetPipelineSummaryParams = zod.object({
  jobId: zod.coerce.number(),
});

export const GetPipelineSummaryResponse = zod.object({
  jobId: zod.number(),
  jobTitle: zod.string(),
  capacity: zod.number(),
  activeCount: zod.number(),
  waitlistCount: zod.number(),
  inactiveCount: zod.number(),
  totalApplications: zod.number(),
  avgTimeToPromotionSeconds: zod.number().nullish(),
  avgAcknowledgmentTimeSeconds: zod.number().nullish(),
  decayEvents: zod.number(),
});

/**
 * @summary Replay pipeline state from audit logs at a given timestamp
 */
export const ReplayPipelineParams = zod.object({
  jobId: zod.coerce.number(),
});

export const ReplayPipelineQueryParams = zod.object({
  asOf: zod.date().optional(),
});

export const ReplayPipelineResponse = zod.object({
  jobId: zod.number(),
  asOf: zod.coerce.date(),
  activeApplicants: zod.array(
    zod.object({
      applicationId: zod.number(),
      applicantId: zod.number(),
      applicantName: zod.string(),
      status: zod.string(),
    }),
  ),
  waitlistApplicants: zod.array(
    zod.object({
      applicationId: zod.number(),
      applicantId: zod.number(),
      applicantName: zod.string(),
      status: zod.string(),
    }),
  ),
  events: zod.array(
    zod.object({
      id: zod.number(),
      applicationId: zod.number(),
      eventType: zod.string(),
      fromStatus: zod.string().nullish(),
      toStatus: zod.string(),
      metadata: zod.record(zod.string(), zod.unknown()).optional(),
      createdAt: zod.coerce.date(),
    }),
  ),
});