# lib/api-zod/src/generated/types/applicationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */
import type { ApplicationResultStatus } from "./applicationResultStatus";

export interface ApplicationResult {
  applicationId: number;
  applicantId: number;
  jobId: number;
  status: ApplicationResultStatus;
  /** @nullable */
  queuePosition?: number | null;
  message?: string;
}