# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */
import type { AuditLogMetadata } from "./auditLogMetadata";

export interface AuditLog {
  id: number;
  applicationId: number;
  eventType: string;
  /** @nullable */
  fromStatus?: string | null;
  toStatus: string;
  metadata?: AuditLogMetadata;
  createdAt: Date;
}