# lib/api-zod/src/generated/types/withdrawBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */

export interface WithdrawBody {
  applicationId: number;
}